+++ src/app/app.component.spec.ts
@@ -7,7 +7,7 @@ import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { Storage } from '@ionic/storage-angular';
-import { SurveyDataService } from './services/survey-data/survey-data.service';
+import { SurveyDataService } from './services/survey-data/data.service';
 import { Router } from '@angular/router';
 import { BarcodeService } from './services/barcode/barcode.service';
 import study_tasks from '../../cypress/fixtures/study_tasks.json';
@@ -54,7 +54,6 @@ describe('AppComponent', () => {
   it('should create the app', () => {
     expect(component).toBeTruthy();
   });
-
 
   it('should test ngOnInIt', async () => {
     const barcodeService = fixture.debugElement.injector.get(BarcodeService);
