+++ src/app/pages/settings/settings.page.ts
@@ -5,7 +5,7 @@ import { NotificationsService } from '../../services/notification/notifications.
 import { Browser } from '@capacitor/browser';
 import * as moment from 'moment';
 import { TranslateConfigService } from '../../translate-config.service';
-import { SurveyDataService } from '../../services/survey-data/survey-data.service';
+import { SurveyDataService } from '../../services/survey-data/data.service';
 import { StorageService } from '../../services/storage/storage.service';
 import { Capacitor } from '@capacitor/core';
