+++ src/app/services/survey-data/data.service.spec.ts
@@ -8,7 +8,7 @@ import {
   HttpClientTestingModule,
 } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
-import { SurveyDataService } from './survey-data.service';
+import { SurveyDataService } from './data.service';
 import study_tasks from '../../../../cypress/fixtures/study_tasks.json';
 import { StorageService } from '../storage/storage.service';
 import { createSpyFromClass, Spy } from 'jasmine-auto-spies';
